requirements.txt
streamlit
pandas
requests
urllib3
beautifulsoup4
newspaper3k
googlesearch-python
dateparser
google-generativeai
